imp: const values, permission validation

+++ app/src/main/java/com/example/madassessment/Constants.java
@@ -7,7 +7,7 @@ public class Constants {
     */
     public static final Double DEFAULT_LAT = 60.169;
     public static final Double DEFAULT_LON = 24.938;
-    public static final Integer DEFAULT_ZOOM = 12;
+    public static final Double DEFAULT_ZOOM = 12.0;
 
     public static final String  NORMAL_MAP = "normal";
     public static final String  CYCLE_MAP = "cycle";
